pyproject.toml
[tool.poetry]
name = "onepdd"
version = "0.1.0"
description = "Puzzle Driven Development assistant for your Gitea repositories"
authors = ["monomonedula <valh@tuta.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
pyyaml = "^6.0.1"
jinja2 = "^3.1.2"
aiohttp = "^3.8.5"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.5.1"
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
